requirements.txt
certifi==2019.3.9
chardet==3.0.4
graphql-core==2.1
idna==2.8
pathlib==1.0.1
pprint==0.1
promise==2.2.1
python-dotenv==0.10.2
requests==2.22.0
Rx==1.6.1
sgqlc==7.0
six==1.12.0
urllib3==1.25.3
